Added equality check for floats (not really).
This method doesn't check if two floats are exactly equal, but rather, checks if they are "close enough" to be equal.
The value of "close enough" is known as epsilon.

This method makes use of Donald Knuth's algorithm on relative epsilon. If the floats you are comparing are huge, say 1,000,000, the epsilon will be small, say 0.01. But if the floats you are comparing are small, say 0.00001, the epsilon will be tiny, say 0.000000001. However, if both floats are very close to zero, then the absolute epsilon would be used instead.