Resizable smiley.
The smiley can be resized by selecting it, and then, while the cursor is still within the bounds of the smiley, hold down ctrl and click on the smiley and move the mouse to the right to increase radius, and move the mouse to the left to decrease radius.

The radius of the smiley is clamped to its minimum and maximum radius.

Resize on mouse move.