Fix: Smiley resizes weirdly if not at origin.
This commit fixes the issue where if the Smiley is not at the origin of the OpenGL space (middle of the window), scaling problems arise, where the smiley's size jumps around when resized.

That is because previously, we only calculate a vector from the cursor's position to the smiley's position (cursorPositionDiff). However, because we are accessing the x-coordinate value of the cursor, if our Smiley is far away from the origin, but near the cursor, the cursorPositionDiff vector would be small, but the x-coordinate value for the mouse cursor is huge, causing an unexpected change in the radius as soon as we begin to resize.

To fix this, we need to keep track of the position of the cursor during the mouse click. This is so that wherever we click, the vector calculated would be large enough, since it is relative to the cursor's position, to prevent the size of the smiley to change haphazardly.