Added implicit conversions from Color3f to Color4f
Color3f can now be implicitly converted to Color4f (the first 3 parameters of Color4f correspond to that of Color3f while the 4th parameter of Color4f will be set to 1.0f).
This should not break methods previously accepting a Color3f, that will now accept Color4f.