Iterations now used for-each loops, where possible
The iterators using for loops (for (auto iterator{vector.begin()})...) have been replaced with for-each loops (for (type el : vector)).

The reason I switch to for-each loops is due to unneeded dereferencing. Dereferencing a pointer is somewhat slow so it is best to avoid it if possible. Because the vector holds a pointer of Smileys, using the iterator with a for loop requires two levels of dereferencing - one with pointer->member and one with (*iterator).

Using for-each loops instead, only requires one level of dereferencing, which is just pointer->member.