Variable to store the current colour of the Smiley
Code in Smiley.cpp has been updated to make use of the new variable. This is not the full transition however, as this is not the master branch. We currently need to do this to allow the poof particles to take on the colour of the Smileys.

The full transition will be performed once the Poof Particles are complete.