Use of size_type when iterating vector.
Because the data type returned from the "capacity" method is "std::vector<...>::size_type", which is an unsigned integral type, comparing it with signed integral types (like int) results in a mismatch of type.

So, we solve this by using a type alias so that we can also initialize our indexer to the right type.